researches/MAI-UI/trainer/requirements.txt
# MAI-UI Training Requirements

# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# RL training
trl>=0.7.0

# Android environment (optional, for RL training)
# android_env

# Data processing
pillow>=10.0.0
pyyaml>=6.0

# Utilities
tqdm>=4.66.0
